fix(tts): X-TTS-Mode header reports the actual backend (not the request)
End-to-end against the 4090 worked but the response header lied:
when qwen3_remote silently fell back to Piper (e.g. a transient
"Connection refused" right after server restart), X-TTS-Mode still
said "qwen3_remote" because the route just echoed the config value.

backend/samantha/tts.py:
- synth() now returns (wav_bytes, backend_used). The caller knows
  exactly which backend produced the audio: "qwen3_remote",
  "piper", or "empty" (no text).
- Renamed legacy state: kept _voice_load_failed naming consistent.

backend/samantha/api.py:
- /speak uses the returned mode_used to set X-TTS-Mode. Observability
  no longer lies on fallback.

tests (60 still passing):
- test_synth_empty_text_returns_empty_bytes asserts the new tuple.
- test_synth_produces_riff_wave pins tts_backend=piper.
- test_qwen3_remote_falls_back_to_piper_on_http_error asserts
  mode_used == "piper" after fallback.

tts-server/server.py:
- Swapped torch_dtype= → dtype= (transformers 4.57+ renamed it;
  old name still works but emits a DeprecationWarning).

End-to-end verified from the Mac backend → 4090 TTS:
  curl /speak  →  X-TTS-Mode: qwen3_remote
                   audio/wav, 24000 Hz mono (Qwen3 native sr)
                   RTF 0.5 steady-state on 4090 fp16 (no flash-attn)

+++ tts-server/server.py
@@ -122,10 +122,13 @@ def get_model():
         f"loading Qwen3-TTS from {MODEL_PATH} (device_map={device}, dtype={dtype})"
     )
     t0 = time.perf_counter()
+    # Note: transformers 4.57+ renamed `torch_dtype` → `dtype` in
+    # from_pretrained. Both still work but `torch_dtype` emits a
+    # DeprecationWarning. Use the new name.
     _model = Qwen3TTSModel.from_pretrained(
         MODEL_PATH,
         device_map=device,
-        torch_dtype=dtype,
+        dtype=dtype,
     )
     logger.info(f"Qwen3-TTS ready in {time.perf_counter() - t0:.1f}s")
     return _model
